+++ astropy_helpers/tests/test_openmp_helpers.py
@@ -3,15 +3,23 @@ import sys
 from copy import deepcopy
 from distutils.core import Extension
 
-from ..openmp_helpers import add_openmp_flags_if_available
+from ..openmp_helpers import add_openmp_flags_if_available, generate_openmp_enabled_py
 from ..setup_helpers import _module_state, register_commands
 
 IS_TRAVIS_LINUX = os.environ.get('TRAVIS_OS_NAME', None) == 'linux'
+IS_TRAVIS_OSX = os.environ.get('TRAVIS_OS_NAME', None) == 'osx'
 IS_APPVEYOR = os.environ.get('APPVEYOR', None) == 'True'
 PY3_LT_35 = sys.version_info[0] == 3 and sys.version_info[1] < 5
 
 _state = None
 
+try:
+    TEST_OPENMP = 'True' == os.environ['TEST_OPENMP']
+except KeyError:
+    if IS_APPVEYOR:
+        TEST_OPENMP = True
+    else:
+        raise
 
 def setup_function(function):
     global state
@@ -34,5 +42,37 @@ def test_add_openmp_flags_if_available():
     # Having this is useful because we'll find out if OpenMP no longer works
     # for any reason on platforms on which it does work at the time of writing.
     # OpenMP doesn't work on Python 3.x where x<5 on AppVeyor though.
-    if IS_TRAVIS_LINUX or (IS_APPVEYOR and not PY3_LT_35):
-        assert using_openmp
+    if IS_TRAVIS_LINUX or IS_TRAVIS_OSX or (IS_APPVEYOR and not PY3_LT_35):
+        if TEST_OPENMP:
+            assert using_openmp
+        else:
+            assert not using_openmp
+
+def test_generate_openmp_enabled_py():
+
+    register_commands('openmp_autogeneration_testing', '0.0', False)
+
+    # Test file generation
+    generate_openmp_enabled_py('')
+    assert os.path.isfile('openmp_enabled.py')
+
+    with open('openmp_enabled.py', 'r') as fid:
+        contents = fid.read()
+    print(contents)
+
+    # Travis OSX tests experience an unstable module import error.
+    # Sometimes finding and importing the module, `openmp_enabled`,
+    # more often not. Adding '.' to the path seems to stabilize the
+    # issue, even though it should not be needed.
+    sys.path.append('.')
+    from openmp_enabled import is_openmp_enabled
+
+    is_openmp_enabled = is_openmp_enabled()
+
+    # Test is_openmp_enabled()
+    assert isinstance(is_openmp_enabled, bool)
+
+    if TEST_OPENMP:
+        assert is_openmp_enabled
+    else:
+        assert not is_openmp_enabled